Add the upcoming Christmas year to default emails
The script will now calculate the year for the upcoming Christmas based
on the local date. If it is before Christmas, you get this year. If it
is Christmas or one of the last six days of December, you get next year.

This year is then passed to the email templates so you can easily
distinguish this year's assignments from previous years.