Refactor assignments with hungarian algorithm
The Hungarian algorithm or Kuhn-Munkres assignment algorithm is a way
of turning a matrix of costs into a set of optimal assignments in
O(n^3) time (which is surprisingly as good as this problem gets). Using
it allows me to replace my previous "retry until you get something
valid" approach. Not only is that more reliable, but it is simpler to
reason about and add future features to.